% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dboots_ci.R
\name{boot_ci}
\alias{boot_ci}
\title{double bootstrap confidence interval}
\usage{
boot_ci(N, rho, alpha)
}
\arguments{
\item{N}{number of replicates}

\item{rho}{correlation estimator}

\item{alpha}{level of significant for rho confidence interval}
}
\value{
None
}
\description{
double bootstrap confidence interval
}
