% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_fun.R
\name{boot_fun}
\alias{boot_fun}
\title{double bootstrap}
\usage{
boot_fun(nboot, xdata = xdata)
}
\arguments{
\item{nboot}{number of boots}

\item{xdata}{sub data set}
}
\value{
None
}
\description{
double bootstrap
}
