% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dboots.R
\name{boot_out}
\alias{boot_out}
\title{double bootstrap output}
\usage{
boot_out(nboot, xdata = xdata)
}
\arguments{
\item{nboot}{number of boots}

\item{xdata}{sub data set}
}
\value{
None
}
\description{
double bootstrap output
}
