% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_post_fun.R
\name{con_post_fun}
\alias{con_post_fun}
\title{conditional posterior}
\usage{
con_post_fun(x, mu_nz, mu_nx, sd_nz, sd_nx, rho_n)
}
\arguments{
\item{x}{delta_pCR}

\item{mu_nz}{mean of z (log hr of EFS)}

\item{mu_nx}{mean of x (delta_pCR)}

\item{sd_nz}{sd of z}

\item{sd_nx}{sd of x}

\item{rho_n}{correlation}
}
\value{
None
}
\description{
conditional posterior
}
