% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gene2.R
\name{data_fun}
\alias{data_fun}
\title{data generation.}
\usage{
data_fun(n, mm, lambda0, pp0, pp1, hr_pcr, mu2_hr)
}
\arguments{
\item{n}{sample size in Phase II}

\item{mm}{additional patients in Phase III}

\item{lambda0}{baseline hazard rate with pCR=No}

\item{pp0}{pCR rate in control arm}

\item{pp1}{pCR rate in trt arm}

\item{hr_pcr}{exp(beta2)}

\item{mu2_hr}{prior for HR mean in Phase III}
}
\value{
None
}
\description{
data generation.
}
