% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main_fun}
\alias{main_fun}
\title{estimation main function}
\usage{
main_fun(
  M,
  nboot,
  alpha,
  n,
  mm,
  lambda0,
  pp0,
  pp1,
  hr_pcr,
  mu1,
  mu2_hr,
  v20,
  rho0,
  cri,
  tau
)
}
\arguments{
\item{M}{number of replicates}

\item{nboot}{number of boots}

\item{alpha}{level of significant for rho confidence interval}

\item{n}{sample size in Phase II}

\item{mm}{additional patients in Phase III}

\item{lambda0}{baseline hazard rate with pCR=No}

\item{pp0}{pCR rate in control arm}

\item{pp1}{pCR rate in trt arm}

\item{hr_pcr}{exp(beta2)}

\item{mu1}{prior for delta pCR mean in Phase II}

\item{mu2_hr}{prior for HR mean in Phase III}

\item{v20}{loghr variance}

\item{rho0}{suppose highly correlated}

\item{cri}{target HR}

\item{tau}{cutoff posterior probability}
}
\value{
None
}
\description{
estimation main function
}
