% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\name{my_stats}
\alias{my_stats}
\title{estimator generation.}
\usage{
my_stats(
  M = 1000,
  n = 300,
  mm = 300,
  lambda0 = 1,
  pp0 = 0.5,
  pp1 = 0.7,
  hr_pcr = 0.24,
  mu2_hr = 0.7
)
}
\arguments{
\item{M}{number of replicates}

\item{n}{sample size in Phase II}

\item{mm}{additional patients in Phase III}

\item{lambda0}{baseline hazard rate with pCR=No}

\item{pp0}{pCR rate in control arm}

\item{pp1}{pCR rate in trt arm}

\item{hr_pcr}{exp(beta2)}

\item{mu2_hr}{prior for HR mean in Phase III}
}
\value{
None
}
\description{
estimator generation.
}
