% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN.R
\name{pCR_est}
\alias{pCR_est}
\title{run the main function}
\usage{
pCR_est(alpha, mu2_hr, cri, tau, nboot)
}
\arguments{
\item{alpha}{significant level}

\item{mu2_hr}{prior of HR mean}

\item{cri}{target HR}

\item{tau}{cutoff posterior probability}

\item{nboot}{number of boots}
}
\value{
None
}
\description{
run the main function
}
