% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_fun.R
\name{post_fun}
\alias{post_fun}
\title{posterior estimator}
\usage{
post_fun(x, mu0, sigma0, smu, ssigma, cri_hr)
}
\arguments{
\item{x}{delta_pCR}

\item{mu0}{prior of mean of z (log hr of EFS)}

\item{sigma0}{priof covariance}

\item{smu}{estimated mu}

\item{ssigma}{covariance matrix}

\item{cri_hr}{target HR}
}
\value{
None
}
\description{
posterior estimator
}
