% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pCR.R
\name{print_pCR}
\alias{print_pCR}
\title{print estimation results}
\usage{
print_pCR(x, ...)
}
\arguments{
\item{x}{estimation output}

\item{...}{other arguments}
}
\value{
None
}
\description{
print estimation results
}
